backend/requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
email-validator==2.1.0.post1
python-dotenv==1.0.1
